+++ packages/alfa-rules/test/common/predicate/is-at-the-start.spec.tsx
@@ -1,72 +1,7 @@
 import { h } from "@siteimprove/alfa-dom/h";
 import { test } from "@siteimprove/alfa-test";
 
-import { None, Option } from "@siteimprove/alfa-option";
-
-import { lowestCommonAncestor } from "../../../src/common/expectation/get-nodes-between";
 import { isAtTheStart } from "../../../src/common/predicate";
-
-test("lowestCommonAncestor() returns the parent of two siblings", (t) => {
-  const node1 = <span>Hello</span>;
-  const node2 = <span>World</span>;
-  const ancestor = (
-    <div>
-      {node1} {node2}
-    </div>
-  );
-
-  t.deepEqual(lowestCommonAncestor(node1, node2), Option.of(ancestor));
-});
-
-test("lowestCommonAncestor() returns the LCA of two nodes at different depths", (t) => {
-  const node1 = <span>Hello</span>;
-  const node2 = <span>World</span>;
-  const ancestor = (
-    <div>
-      <div>
-        <div>{node1}</div>
-      </div>{" "}
-      <div>{node2}</div>
-    </div>
-  );
-
-  h.document([
-    <div>
-      <div>ancestor</div>
-    </div>,
-  ]);
-
-  t.deepEqual(lowestCommonAncestor(node1, node2), Option.of(ancestor));
-});
-
-test("lowestCommonAncestor() returns one node when it is ancestor of the other", (t) => {
-  const node1 = <span>Hello</span>;
-  const ancestor = (
-    <div>
-      <div>
-        <div>{node1}</div>
-      </div>
-    </div>
-  );
-
-  h.document([
-    <div>
-      <div>ancestor</div>
-    </div>,
-  ]);
-
-  t.deepEqual(lowestCommonAncestor(node1, ancestor), Option.of(ancestor));
-});
-
-test("lowestCommonAncestor() returns None when nodes are in different trees", (t) => {
-  const node1 = <span>Hello</span>;
-  const node2 = <span>World</span>;
-
-  h.document([<div>{node1}</div>]);
-  h.document([<div>{node2}</div>]);
-
-  t.deepEqual(lowestCommonAncestor(node1, node2), None);
-});
 
 test("isAtTheStart() returns true on an element itself", (t) => {
   const target = <div />;
