+++ packages/alfa-rules/tsconfig.json
@@ -15,6 +15,7 @@
     "src/common/expectation/get-colors.ts",
     "src/common/expectation/get-refresh-time.ts",
     "src/common/expectation/is-passed.ts",
+    "src/common/expectation/lowest-common-ancestor.ts",
     "src/common/expectation/media-text-alternative.ts",
     "src/common/expectation/video-description-track-accurate.ts",
     "src/common/group.ts",
@@ -156,10 +157,11 @@
     "test/common/evaluate.ts",
     "test/common/oracle.ts",
     "test/common/outcome.ts",
+    "test/common/expectation/get-colors.spec.tsx",
+    "test/common/expectation/lowest-common-ancestor.spec.tsx",
     "test/common/predicate/is-at-the-start.spec.tsx",
     "test/common/predicate/is-visible.spec.tsx",
     "test/common/predicate/has-role.spec.tsx",
-    "test/common/expectation/get-colors.spec.tsx",
     "test/sia-r1/rule.spec.tsx",
     "test/sia-r2/rule.spec.tsx",
     "test/sia-r3/rule.spec.tsx",
